requirements.txt
 google-cloud-speech>=2.0.0
 pyaudio